Transaction model manage account as object

+++ lib/app/modules/finance/account/models/account_model.dart
@@ -79,8 +79,8 @@ class Account {
         name: map["name"],
         id: map["id"],
         photoURL: map["photoURL"],
-        balanceCredit: map['balanceCredit'],
-        balanceDebit: map['balanceDebit']);
+        balanceCredit: map['balanceCredit'].toDouble(),
+        balanceDebit: map['balanceDebit'].toDouble());
   }
 
   Map<String, dynamic> toMap() {
